Various (fairly large) improvements
Attempted to fix a crash bug involving pathfinding (failed, but did add some safeguards and tidied up some code).
Overhauled the player movement to use a Vector-based system with lots of comments. This is more compact, easier to edit, and avoids repeated code. It also means the velocity is normalised, which is always nice. I made the speed boost work like the slowed debuff (they can now cancel each other out, and it's much simpler) and made the confusion affect the input only, rather than the velocity directly. tl;dr - the movement and powerup systems are now a bit more modular and consistent, with nice comments to boot.
Finally, re-ordered some of the player's code to be more readable.